zero-base the chat page
